+++ CHANGES.rst
@@ -190,6 +190,10 @@ Bug Fixes
 
 - ``astropy.table``
 
+  - Assigning a logarithmic unit to a ``QTable`` column that did not have a
+    unit yet now correctly turns it into the appropriate function quantity
+    subclass (such as ``Magnitude`` or ``Dex``). [#5345]
+
 - ``astropy.time``
 
 - ``astropy.units``
@@ -208,6 +212,9 @@ Bug Fixes
   - Conversion from quantities to logarithmic units now correctly causes a
     logarithmic quantity such as ``Magnitude`` to be returned. [#5183]
   
+
+  - Ensure ``!=`` also works for function units such as ``MagUnit`` [#5345]
+
 - ``astropy.utils``
 
 - ``astropy.visualization``